created separate file for data conversion in promql and sql.
