+++ web/src/utils/dashboard/convertPromQLData.ts
@@ -13,6 +13,7 @@
 //  limitations under the License.
 
 import moment from "moment";
+import { formatDate, formatUnitValue, getUnitValue } from "./convertDataIntoUnitValue";
 
 /**
  * Converts PromQL data into a format suitable for rendering a chart.
@@ -311,137 +312,6 @@ const getLegendPosition = (legendPosition: string) => {
   }
 };
 
-/**
- * Retrieves the value and unit of a given value based on the specified unit and custom unit.
- *
- * @param {any} value - The value to be converted.
- * @param {string} unit - The unit of the value.
- * @param {string} customUnit - The custom unit to be used if the unit is 'custom'.
- * @return {object} An object containing the converted value and its corresponding unit.
- */
-const getUnitValue = (value: any, unit: string, customUnit: string) => {
-
-  switch (unit) {
-    case "bytes": {
-      const units = ["B", "KB", "MB", "GB", "TB"];
-      for (let unit of units) {
-        if (value < 1024) {
-          return {
-            value: `${parseFloat(value).toFixed(2)}`,
-            unit: `${unit}`,
-          };
-        }
-        value /= 1024;
-      }
-      return {
-        value: `${parseFloat(value).toFixed(2)}`,
-        unit: "PB",
-      };
-    }
-    case "seconds": {
-      const units = [
-        { unit: "ms", divisor: 0.001 },
-        { unit: "s", divisor: 1 },
-        { unit: "m", divisor: 60 },
-        { unit: "h", divisor: 3600 },
-        { unit: "D", divisor: 86400 },
-        { unit: "M", divisor: 2592000 }, // Assuming 30 days in a month
-        { unit: "Y", divisor: 31536000 }, // Assuming 365 days in a year
-      ];
-      for (const unitInfo of units) {
-        const unitValue = value ? value / unitInfo.divisor : 0;
-        if (unitValue >= 1 && unitValue < 1000) {
-          return {
-            value: unitValue.toFixed(2),
-            unit: unitInfo.unit,
-          };
-        }
-      }
-
-      // If the value is too large to fit in any unit, return the original seconds
-      return {
-        value: value,
-        unit: "s",
-      };
-    }
-    case "bps": {
-      const units = ["B", "KB", "MB", "GB", "TB"];
-      for (let unit of units) {
-        if (value < 1024) {
-          return {
-            value: `${parseFloat(value).toFixed(2)}`,
-            unit: `${unit}/s`,
-          };
-        }
-        value /= 1024;
-      }
-      return {
-        value: `${parseFloat(value).toFixed(2)}`,
-        unit: "PB/s",
-      };
-    }
-    case "percent-1": {
-      return {
-        value: `${(parseFloat(value) * 100).toFixed(2)}`,
-        unit: "%",
-      };
-      // `${parseFloat(value) * 100}`;
-    }
-    case "percent": {
-      return {
-        value: `${parseFloat(value).toFixed(2)}`,
-        unit: "%",
-      };
-      // ${parseFloat(value)}`;
-    }
-    case "custom": {
-      return {
-        value: `${value}`,
-        unit: `${customUnit ?? ""}`,
-      };
-    }
-    case "default": {
-      return {
-        value: value,
-        unit: "",
-      };
-    }
-    default: {
-      return {
-        value: value,
-        unit: "",
-      };
-    }
-  }
-};
-
-/**
- * Formats a unit value by concatenating the value and unit.
- *
- * @param {any} obj - The object containing the value and unit.
- * @return {string} The formatted unit value.
- */
-const formatUnitValue = (obj: any) => {
-  return `${obj.value}${obj.unit}`;
-};
-
-/**
- * Formats a given date into a string representation.
- * Useful for showing date in a timeseries chart
- *
- * @param {any} date - The date to be formatted.
- * @return {string} The formatted date string
- */
-const formatDate = (date: any) => {
-  const day = String(date.getDate()).padStart(2, "0");
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const year = String(date.getFullYear()).slice(2);
-  const hours = String(date.getHours()).padStart(2, "0");
-  const minutes = String(date.getMinutes()).padStart(2, "0");
-  const seconds = String(date.getSeconds()).padStart(2, "0");
-
-  return `${day}-${month}-${year} ${hours}:${minutes}:${seconds}`;
-};
 
 /**
  * Returns the props object based on the given chart type.
